+++ src/lib/variable_data.ts
@@ -1,7 +1,7 @@
 import type {Array_Element} from '@ryanatkn/belt/types.js';
 
-export type Size_Variant = Array_Element<typeof size_variants>;
-export const size_variants = [
+export type Size_Variant = Array_Element<typeof font_size_variants>;
+export const font_size_variants = [
 	'xs',
 	'sm',
 	'md',
@@ -17,10 +17,17 @@ export const size_variants = [
 	'xl9',
 ] as const;
 
-export const size_names = size_variants.map((s) => 'size_' + s);
+export const font_size_names = font_size_variants.map((s) => 'font_size_' + s);
 
-export type Text_Variant = Array_Element<typeof text_variants>;
-export const text_variants = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10] as const;
+export type Font_Family_Variant = Array_Element<typeof font_family_variants>;
+export const font_family_variants = [
+	'font_family_sans',
+	'font_family_serif',
+	'font_family_mono',
+] as const;
+
+export type Text_Color_Variant = Array_Element<typeof text_color_variants>;
+export const text_color_variants = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10] as const;
 
 export type Space_Variant = Array_Element<typeof space_variants>;
 export const space_variants = [
@@ -52,8 +59,8 @@ export const space_variants = [
 export type Distance_Variant = Array_Element<typeof distance_variants>;
 export const distance_variants = ['xs', 'sm', 'md', 'lg', 'xl'] as const;
 
-export type Radius_Variant = Array_Element<typeof radius_variants>;
-export const radius_variants = ['xs3', 'xs2', 'xs', 'sm', 'md', 'lg', 'xl'] as const;
+export type Border_Radius_Variant = Array_Element<typeof border_radius_variants>;
+export const border_radius_variants = ['xs3', 'xs2', 'xs', 'sm', 'md', 'lg', 'xl'] as const;
 
 export type Line_Height_Variant = Array_Element<typeof line_height_variants>;
 export const line_height_variants = ['xs', 'sm', 'md', 'lg', 'xl'] as const;
@@ -69,8 +76,8 @@ export const shadow_variant_prefixes = [
 	'shadow_inset_bottom_',
 ] as const;
 
-export type Shadow_Size_Variant = Array_Element<typeof shadow_size_variants>;
-export const shadow_size_variants = ['xs', 'sm', 'md', 'lg', 'xl'] as const;
+export type Shadow_Size_Variant = Array_Element<typeof shadow_font_size_variants>;
+export const shadow_font_size_variants = ['xs', 'sm', 'md', 'lg', 'xl'] as const;
 
 export type Shadow_Alpha_Variant = Array_Element<typeof shadow_alpha_variants>;
 export const shadow_alpha_variants = [1, 2, 3, 4, 5] as const;