+++ src/lib/variables.ts
@@ -706,41 +706,42 @@ export const text_active: Style_Variable = {
 };
 
 /* fonts */
+export const font_family_sans: Style_Variable = {
+	name: 'font_family_sans',
+	light: "system-ui, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji'",
+	summary:
+		'@source https://kilianvalkhof.com/2022/css-html/your-css-reset-needs-text-size-adjust-probably/',
+};
+export const font_family_serif: Style_Variable = {
+	name: 'font_family_serif',
+	light: "'DM Serif Display', Georgia, serif",
+	summary: '@source https://svelte.dev/',
+};
+export const font_family_mono: Style_Variable = {
+	name: 'font_family_mono',
+	light: "ui-monospace, 'Fira Mono', monospace",
+};
+
+/* sizes like font-size */
+export const font_size_xs: Style_Variable = {name: 'font_size_xs', light: '1rem'};
+export const font_size_sm: Style_Variable = {name: 'font_size_sm', light: '1.3rem'};
+export const font_size_md: Style_Variable = {name: 'font_size_md', light: '1.6rem'};
+export const font_size_lg: Style_Variable = {name: 'font_size_lg', light: '2.04rem'};
+export const font_size_xl: Style_Variable = {name: 'font_size_xl', light: '2.59rem'};
+export const font_size_xl2: Style_Variable = {name: 'font_size_xl2', light: '3.29rem'};
+export const font_size_xl3: Style_Variable = {name: 'font_size_xl3', light: '4.19rem'};
+export const font_size_xl4: Style_Variable = {name: 'font_size_xl4', light: '5.33rem'};
+export const font_size_xl5: Style_Variable = {name: 'font_size_xl5', light: '6.78rem'};
+export const font_size_xl6: Style_Variable = {name: 'font_size_xl6', light: '8.62rem'};
+export const font_size_xl7: Style_Variable = {name: 'font_size_xl7', light: '10.97rem'};
+export const font_size_xl8: Style_Variable = {name: 'font_size_xl8', light: '13.95rem'};
+export const font_size_xl9: Style_Variable = {name: 'font_size_xl9', light: '17.74rem'};
+
 export const line_height_xs: Style_Variable = {name: 'line_height_xs', light: '1'};
 export const line_height_sm: Style_Variable = {name: 'line_height_sm', light: '1.2'};
 export const line_height_md: Style_Variable = {name: 'line_height_md', light: '1.4'};
 export const line_height_lg: Style_Variable = {name: 'line_height_lg', light: '1.6'};
 export const line_height_xl: Style_Variable = {name: 'line_height_xl', light: '2'};
-export const font_sans: Style_Variable = {
-	name: 'font_sans',
-	light: "system-ui, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji'",
-	summary:
-		'@source https://kilianvalkhof.com/2022/css-html/your-css-reset-needs-text-size-adjust-probably/',
-};
-export const font_serif: Style_Variable = {
-	name: 'font_serif',
-	light: "'DM Serif Display', Georgia, serif",
-	summary: '@source https://svelte.dev/',
-};
-export const font_mono: Style_Variable = {
-	name: 'font_mono',
-	light: "ui-monospace, 'Fira Mono', monospace",
-};
-
-/* sizes like font-size */
-export const size_xs: Style_Variable = {name: 'size_xs', light: '1rem'};
-export const size_sm: Style_Variable = {name: 'size_sm', light: '1.3rem'};
-export const size_md: Style_Variable = {name: 'size_md', light: '1.6rem'};
-export const size_lg: Style_Variable = {name: 'size_lg', light: '2.04rem'};
-export const size_xl: Style_Variable = {name: 'size_xl', light: '2.59rem'};
-export const size_xl2: Style_Variable = {name: 'size_xl2', light: '3.29rem'};
-export const size_xl3: Style_Variable = {name: 'size_xl3', light: '4.19rem'};
-export const size_xl4: Style_Variable = {name: 'size_xl4', light: '5.33rem'};
-export const size_xl5: Style_Variable = {name: 'size_xl5', light: '6.78rem'};
-export const size_xl6: Style_Variable = {name: 'size_xl6', light: '8.62rem'};
-export const size_xl7: Style_Variable = {name: 'size_xl7', light: '10.97rem'};
-export const size_xl8: Style_Variable = {name: 'size_xl8', light: '13.95rem'};
-export const size_xl9: Style_Variable = {name: 'size_xl9', light: '17.74rem'};
 
 /* links */
 export const link_color: Style_Variable = {
@@ -868,6 +869,9 @@ export const border_color_i: Style_Variable = {
 	light: 'var(--color_i_5)',
 };
 export const border_width: Style_Variable = {name: 'border_width', light: 'var(--border_width_1)'};
+// These use numbers instead of named size variants because
+// they more directly map to how I think about border widths.
+// But maye this could be expanded/rethought.
 export const border_width_1: Style_Variable = {name: 'border_width_1', light: '1px'};
 export const border_width_2: Style_Variable = {name: 'border_width_2', light: '2px'};
 export const border_width_3: Style_Variable = {name: 'border_width_3', light: '3px'};
@@ -898,13 +902,13 @@ export const outline_color: Style_Variable = {
 };
 
 /* border radii */
-export const radius_xl: Style_Variable = {name: 'radius_xl', light: '5.5rem'};
-export const radius_lg: Style_Variable = {name: 'radius_lg', light: '3.4rem'};
-export const radius_md: Style_Variable = {name: 'radius_md', light: '2.1rem'};
-export const radius_sm: Style_Variable = {name: 'radius_sm', light: '1.3rem'};
-export const radius_xs: Style_Variable = {name: 'radius_xs', light: '0.8rem'};
-export const radius_xs2: Style_Variable = {name: 'radius_xs2', light: '0.5rem'};
-export const radius_xs3: Style_Variable = {name: 'radius_xs3', light: '0.3rem'};
+export const border_radius_xs3: Style_Variable = {name: 'border_radius_xs3', light: '0.3rem'};
+export const border_radius_xs2: Style_Variable = {name: 'border_radius_xs2', light: '0.5rem'};
+export const border_radius_xs: Style_Variable = {name: 'border_radius_xs', light: '0.8rem'};
+export const border_radius_sm: Style_Variable = {name: 'border_radius_sm', light: '1.3rem'};
+export const border_radius_md: Style_Variable = {name: 'border_radius_md', light: '2.1rem'};
+export const border_radius_lg: Style_Variable = {name: 'border_radius_lg', light: '3.4rem'};
+export const border_radius_xl: Style_Variable = {name: 'border_radius_xl', light: '5.5rem'};
 
 /* button shadows */
 export const button_shadow: Style_Variable = {
@@ -1162,11 +1166,26 @@ export const shadow_alpha_5: Style_Variable = {
 /* icons */
 /* these decrease by the golden ratio, rounded to the nearest pixel,
 	and they're insensitive to font size (`px` not `rem`) */
-export const icon_size_xs: Style_Variable = {name: 'icon_size_xs', light: icon_sizes.icon_size_xs};
-export const icon_size_sm: Style_Variable = {name: 'icon_size_sm', light: icon_sizes.icon_size_sm};
-export const icon_size_md: Style_Variable = {name: 'icon_size_md', light: icon_sizes.icon_size_md};
-export const icon_size_lg: Style_Variable = {name: 'icon_size_lg', light: icon_sizes.icon_size_lg};
-export const icon_size_xl: Style_Variable = {name: 'icon_size_xl', light: icon_sizes.icon_size_xl};
+export const icon_size_xs: Style_Variable = {
+	name: 'icon_size_xs',
+	light: icon_sizes.icon_size_xs,
+};
+export const icon_size_sm: Style_Variable = {
+	name: 'icon_size_sm',
+	light: icon_sizes.icon_size_sm,
+};
+export const icon_size_md: Style_Variable = {
+	name: 'icon_size_md',
+	light: icon_sizes.icon_size_md,
+};
+export const icon_size_lg: Style_Variable = {
+	name: 'icon_size_lg',
+	light: icon_sizes.icon_size_lg,
+};
+export const icon_size_xl: Style_Variable = {
+	name: 'icon_size_xl',
+	light: icon_sizes.icon_size_xl,
+};
 export const icon_size_xl2: Style_Variable = {
 	name: 'icon_size_xl2',
 	light: icon_sizes.icon_size_xl2,
@@ -1381,29 +1400,30 @@ export const default_variables: Array<Style_Variable> = [
 	text_active,
 
 	/* fonts */
+	font_family_sans,
+	font_family_serif,
+	font_family_mono,
+
+	/* font size */
+	font_size_xs,
+	font_size_sm,
+	font_size_md,
+	font_size_lg,
+	font_size_xl,
+	font_size_xl2,
+	font_size_xl3,
+	font_size_xl4,
+	font_size_xl5,
+	font_size_xl6,
+	font_size_xl7,
+	font_size_xl8,
+	font_size_xl9,
+
 	line_height_xs,
 	line_height_sm,
 	line_height_md,
 	line_height_lg,
 	line_height_xl,
-	font_sans,
-	font_serif,
-	font_mono,
-
-	/* sizes like font-size */
-	size_xs,
-	size_sm,
-	size_md,
-	size_lg,
-	size_xl,
-	size_xl2,
-	size_xl3,
-	size_xl4,
-	size_xl5,
-	size_xl6,
-	size_xl7,
-	size_xl8,
-	size_xl9,
 
 	/* links */
 	link_color,
@@ -1477,13 +1497,13 @@ export const default_variables: Array<Style_Variable> = [
 	outline_color,
 
 	/* border radii */
-	radius_xl,
-	radius_lg,
-	radius_md,
-	radius_sm,
-	radius_xs,
-	radius_xs2,
-	radius_xs3,
+	border_radius_xs3,
+	border_radius_xs2,
+	border_radius_xs,
+	border_radius_sm,
+	border_radius_md,
+	border_radius_lg,
+	border_radius_xl,
 
 	/* button styles */
 	button_shadow,