+++ src/routes/docs/layout/+page.svelte
@@ -80,6 +80,6 @@
 
 	.computed_value {
 		margin-left: var(--space_md);
-		font-family: var(--font_mono);
+		font-family: var(--font_family_mono);
 	}
 </style>