+++ src/routes/docs/shadows/+page.svelte
@@ -8,7 +8,7 @@
 	import Style_Variable_Button from '$routes/Style_Variable_Button.svelte';
 	import {
 		color_variants,
-		shadow_size_variants,
+		shadow_font_size_variants,
 		shadow_variant_prefixes,
 		shadow_alpha_variants,
 		type Shadow_Size_Variant,
@@ -41,13 +41,13 @@
 		>
 		{@render shadow_example_header()}
 		{#each shadow_variant_prefixes as shadow_variant_prefix (shadow_variant_prefix)}
-			{#each shadow_size_variants as shadow_size_variant (shadow_size_variant)}
+			{#each shadow_font_size_variants as shadow_font_size_variant (shadow_font_size_variant)}
 				<div class="shadow_example">
-					<div class="shadow_main_example {shadow_variant_prefix}{shadow_size_variant}">
-						<Style_Variable_Button name="{shadow_variant_prefix}{shadow_size_variant}" />
+					<div class="shadow_main_example {shadow_variant_prefix}{shadow_font_size_variant}">
+						<Style_Variable_Button name="{shadow_variant_prefix}{shadow_font_size_variant}" />
 						<Style_Variable_Button name="shadow_color" />
 					</div>
-					{@render shadow_variant_examples(null, shadow_size_variant, shadow_variant_prefix)}
+					{@render shadow_variant_examples(null, shadow_font_size_variant, shadow_variant_prefix)}
 				</div>
 			{/each}
 		{/each}
@@ -61,17 +61,17 @@
 		<div class="panel fg_3 p_md">
 			{@render shadow_example_header()}
 			{#each shadow_variant_prefixes as shadow_variant_prefix (shadow_variant_prefix)}
-				{#each shadow_size_variants as shadow_size_variant (shadow_size_variant)}
+				{#each shadow_font_size_variants as shadow_font_size_variant (shadow_font_size_variant)}
 					<div class="shadow_example">
 						<div
-							class="shadow_main_example {shadow_variant_prefix}{shadow_size_variant} shadow_color_highlight"
+							class="shadow_main_example {shadow_variant_prefix}{shadow_font_size_variant} shadow_color_highlight"
 						>
-							<Style_Variable_Button name="{shadow_variant_prefix}{shadow_size_variant}" />
+							<Style_Variable_Button name="{shadow_variant_prefix}{shadow_font_size_variant}" />
 							<Style_Variable_Button name="shadow_color_highlight" />
 						</div>
 						{@render shadow_variant_examples(
 							'highlight',
-							shadow_size_variant,
+							shadow_font_size_variant,
 							shadow_variant_prefix,
 						)}
 					</div>
@@ -88,15 +88,19 @@
 		<div class="panel darken_3 p_md">
 			{@render shadow_example_header()}
 			{#each shadow_variant_prefixes as shadow_variant_prefix (shadow_variant_prefix)}
-				{#each shadow_size_variants as shadow_size_variant (shadow_size_variant)}
+				{#each shadow_font_size_variants as shadow_font_size_variant (shadow_font_size_variant)}
 					<div class="shadow_example">
 						<div
-							class="shadow_main_example {shadow_variant_prefix}{shadow_size_variant} shadow_color_glow"
+							class="shadow_main_example {shadow_variant_prefix}{shadow_font_size_variant} shadow_color_glow"
 						>
-							<Style_Variable_Button name="{shadow_variant_prefix}{shadow_size_variant}" />
+							<Style_Variable_Button name="{shadow_variant_prefix}{shadow_font_size_variant}" />
 							<Style_Variable_Button name="shadow_color_glow" />
 						</div>
-						{@render shadow_variant_examples('glow', shadow_size_variant, shadow_variant_prefix)}
+						{@render shadow_variant_examples(
+							'glow',
+							shadow_font_size_variant,
+							shadow_variant_prefix,
+						)}
 					</div>
 				{/each}
 			{/each}
@@ -111,15 +115,19 @@
 		<div class="panel lighten_3 p_md">
 			{@render shadow_example_header()}
 			{#each shadow_variant_prefixes as shadow_variant_prefix (shadow_variant_prefix)}
-				{#each shadow_size_variants as shadow_size_variant (shadow_size_variant)}
+				{#each shadow_font_size_variants as shadow_font_size_variant (shadow_font_size_variant)}
 					<div class="shadow_example">
 						<div
-							class="shadow_main_example {shadow_variant_prefix}{shadow_size_variant} shadow_color_shroud"
+							class="shadow_main_example {shadow_variant_prefix}{shadow_font_size_variant} shadow_color_shroud"
 						>
-							<Style_Variable_Button name="{shadow_variant_prefix}{shadow_size_variant}" />
+							<Style_Variable_Button name="{shadow_variant_prefix}{shadow_font_size_variant}" />
 							<Style_Variable_Button name="shadow_color_shroud" />
 						</div>
-						{@render shadow_variant_examples('shroud', shadow_size_variant, shadow_variant_prefix)}
+						{@render shadow_variant_examples(
+							'shroud',
+							shadow_font_size_variant,
+							shadow_variant_prefix,
+						)}
 					</div>
 				{/each}
 			{/each}
@@ -136,20 +144,20 @@
 			{@const classes = 'color_' + color_variant}
 			<section>
 				{#each shadow_variant_prefixes as shadow_variant_prefix (shadow_variant_prefix)}
-					{#each shadow_size_variants as shadow_size_variant (shadow_size_variant)}
+					{#each shadow_font_size_variants as shadow_font_size_variant (shadow_font_size_variant)}
 						<div class="shadow_example">
 							<div
-								class="shadow_main_example {shadow_variant_prefix}{shadow_size_variant} shadow_color_{color_variant}"
+								class="shadow_main_example {shadow_variant_prefix}{shadow_font_size_variant} shadow_color_{color_variant}"
 							>
 								<Style_Variable_Button
-									name="{shadow_variant_prefix}{shadow_size_variant}"
+									name="{shadow_variant_prefix}{shadow_font_size_variant}"
 									{classes}
 								/>
 								<Style_Variable_Button name="shadow_color_{color_variant}" {classes} />
 							</div>
 							{@render shadow_variant_examples(
 								color_variant,
-								shadow_size_variant,
+								shadow_font_size_variant,
 								shadow_variant_prefix,
 							)}
 						</div>
@@ -178,12 +186,12 @@
 
 {#snippet shadow_variant_examples(
 	color_variant: Color_Variant | 'highlight' | 'glow' | 'shroud' | null,
-	shadow_size_variant: Shadow_Size_Variant,
+	shadow_font_size_variant: Shadow_Size_Variant,
 	shadow_variant_prefix: string,
 )}
 	<div class="row gap_lg">
 		{#each shadow_alpha_variants as alpha (alpha)}
-			{@const shadow_size = shadow_variant_prefix + shadow_size_variant}
+			{@const shadow_size = shadow_variant_prefix + shadow_font_size_variant}
 			{@const shadow_color = `shadow_color${color_variant ? '_' + color_variant : ''}`}
 			<div
 				title="{shadow_size} with {shadow_color}"
@@ -197,8 +205,8 @@
 	.shadow_example {
 		position: relative;
 		padding: var(--space_md);
-		font-family: var(--font_mono);
-		border-radius: var(--radius_xs3);
+		font-family: var(--font_family_mono);
+		border-radius: var(--border_radius_xs3);
 		display: flex;
 		align-items: center;
 		justify-content: end;